# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation Laravel App Abdul Goffar",
        "version": "1.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "product"
                ],
                "operationId": "get all product",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "description": "current_page",
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "string",
                                            "example": "[]"
                                        },
                                        "first_page_url": {
                                            "description": "first_page_url",
                                            "type": "string",
                                            "example": "http:\\/\\/127.0.0.1:8000\\/api\\/categories?page=1"
                                        },
                                        "from": {
                                            "description": "from",
                                            "type": "string",
                                            "example": "null"
                                        },
                                        "last_page": {
                                            "description": "last_page",
                                            "type": "integer",
                                            "example": "3"
                                        },
                                        "last_page_url": {
                                            "description": "last_page_url",
                                            "type": "string",
                                            "example": "http:\\/\\/127.0.0.1:8000\\/api\\/categories?page=3"
                                        },
                                        "links": {
                                            "description": "links",
                                            "type": "string",
                                            "example": "[]"
                                        },
                                        "next_page_url": {
                                            "description": "next_page_url",
                                            "type": "string",
                                            "example": "null"
                                        },
                                        "path": {
                                            "description": "path",
                                            "type": "string",
                                            "example": "null"
                                        },
                                        "per_page": {
                                            "description": "per_page",
                                            "type": "integer",
                                            "example": "4"
                                        },
                                        "prev_page_url": {
                                            "description": "prev_page_url",
                                            "type": "string",
                                            "example": "null"
                                        },
                                        "to": {
                                            "description": "to",
                                            "type": "string",
                                            "example": "null"
                                        },
                                        "total": {
                                            "description": "total",
                                            "type": "integer",
                                            "example": "7"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuthP": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "product"
                ],
                "operationId": "post data product",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "name",
                                        "type": "string",
                                        "example": "jacket laki-laki"
                                    },
                                    "category": {
                                        "description": "category",
                                        "type": "integer",
                                        "example": "12"
                                    },
                                    "price": {
                                        "description": "price",
                                        "type": "integer",
                                        "example": "120000"
                                    },
                                    "stock": {
                                        "description": "stock",
                                        "type": "integer",
                                        "example": "13452"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "description": "name",
                                            "type": "string",
                                            "example": "jacket laki-laki"
                                        },
                                        "category": {
                                            "description": "category",
                                            "type": "integer",
                                            "example": "12"
                                        },
                                        "price": {
                                            "description": "price",
                                            "type": "integer",
                                            "example": "120000"
                                        },
                                        "stock": {
                                            "description": "stock",
                                            "type": "integer",
                                            "example": "13452"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Data Invalid"
                    }
                },
                "security": [
                    {
                        "bearerAuthP": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": " Use /auth to get the JWT token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "bearerAuthC": {
                "type": "http",
                "description": " Use /auth to get the JWT token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "bearerAuthP": {
                "type": "http",
                "description": " Use /auth to get the JWT token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "bearerAuthU": {
                "type": "http",
                "description": " Use /auth to get the JWT token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}